requirements.txt
cmake
dlib
face_recognition
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models
numpy
Pillow
opencv-python
python-dotenv
